Question 9
I was able to implement a Map to define each day of the week as a number
and then I was able to convert the DateTime provided into a string as
the day of the week (i.e. Monday) and use that as a Key to the map to
then return the Integer as a casted string.