Question 17
Created a List to store all the new Person objects. Then I used a for
each loop to loop through the provided List of Names and create a new
Person for each name all with the same age. I then added each new person
to the List to store the Person and when the loop was finished return
the new List of Person objects.